Allow only long or double constants

+++ src-tests/com/maccasoft/propeller/spin1/Spin1ClockModeTest.java
@@ -46,8 +46,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_0_0_00_000, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(12_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_0_0_00_000, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(12_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -57,8 +57,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_0_0_00_000, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(12_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_0_0_00_000, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(12_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -68,8 +68,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_0_0_00_001, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(20_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_0_0_00_001, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(20_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -80,8 +80,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_0_1_00_010, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_0_1_00_010, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -92,8 +92,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_00_011, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_00_011, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -104,8 +104,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_00_100, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(10_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_00_100, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(10_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -116,8 +116,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_00_101, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(20_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_00_101, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(20_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -128,8 +128,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_00_110, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(40_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_00_110, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(40_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -140,8 +140,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_00_111, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(80_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_00_111, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(80_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -152,8 +152,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_01_011, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_01_011, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -164,8 +164,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_01_100, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(10_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_01_100, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(10_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -176,8 +176,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_01_101, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(20_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_01_101, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(20_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -188,8 +188,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_01_110, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(40_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_01_110, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(40_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -200,8 +200,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_01_111, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(80_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_01_111, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(80_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -212,8 +212,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_10_011, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_10_011, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -224,8 +224,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_10_100, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(10_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_10_100, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(10_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -236,8 +236,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_10_101, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(20_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_10_101, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(20_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -248,8 +248,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_10_110, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(40_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_10_110, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(40_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -260,8 +260,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_10_111, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(80_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_10_111, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(80_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -272,8 +272,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_11_011, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_11_011, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(5_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -284,8 +284,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_11_100, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(10_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_11_100, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(10_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -296,8 +296,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_11_101, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(20_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_11_101, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(20_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -308,8 +308,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_11_110, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(40_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_11_110, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(40_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
     @Test
@@ -320,8 +320,8 @@ class Spin1ClockModeTest {
 
         subject.determineClock();
 
-        Assertions.assertEquals(0b0_1_1_11_111, subject.scope.getSymbol("CLKMODE").getNumber());
-        Assertions.assertEquals(80_000_000, subject.scope.getSymbol("CLKFREQ").getNumber());
+        Assertions.assertEquals(0b0_1_1_11_111, subject.scope.getSymbol("CLKMODE").getNumber().intValue());
+        Assertions.assertEquals(80_000_000, subject.scope.getSymbol("CLKFREQ").getNumber().intValue());
     }
 
 }
